requirements.txt
et-xmlfile==1.1.0
lxml==4.9.3
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-docx==1.0.1
pytz==2023.3.post1
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
